Create Groceries_to_buy_list
In this case, we have a to buy list for your personal groceries. However, there are some problems I haven't been able to figure out yet. These are the following:
Once you log out you lose the entire list and you can't access that list again.
Also, it is not possible to close the program without loosing the list 
-__-